Add support for a custom Scalar type, using templates
Types and classes are now all templated on the scalar type, and a
basic type-traits struct is introduced to carry information about
a given scalar type.

This change has been introduced to support the usage of Automatic
Differentiation tools based on operator overloading of custom
scalar types.

However, the code should be 100% backwards compatible as the iit::rbd
namespace contains explicit instantiations of the templates with the
default scalar 'double', reproducing the status of the namespace prior
to this commit.

The changes include a new header in robcogen/ for the interoperability
with CppAD, defining the ScalarTraits for CppAD types. Also, a simple
binary to test the scalar traits was added.