Rewrite core typedefs and improve {linear,angular}Part()
Define the individual templated types (e.g. 'Mat33<S>') and
then define the inner types in 'Core<S>' in terms of the
firsts, and not vice-versa (as it used to be).

This way, the definition of the individual types is not
a dependent name (like 'Core<S>::Matrix33'), which allows
for inference of the scalar type in templated functions
that take arguments of such types.

Leveraging this improvement, I could have a single,
scalar-templated implementation of 'linearPart()' and
'angularPart()', removing duplication (the templated
version in 'Core<S>' and the one specialized for double).

In the process, some 'typedef' replaced with 'using'.