Fix lack of explicit scalar type argument in internal calls
I was unconditionally calling the 'double' version of
'linearPart()' and 'angularPart()', preventing the
'CTransformCore' to work with custom scalar types.

Note that the bug was NOT induced by the previous commit,
nor the previous commit was strictly necessary to fix this.

In fact, enabling deduction of the scalar type, in
the previous commit, automatically fixes bugs similar
to this one. However, in this case template type
deduction does not work because the calls to the
fuctions do not have arguments like 'Vec<S>' (which
would allow the compiler to infer S), but arguments
of a generic type (which is there for the calling
function to accept matrix expressions).